Updated code after the FIRST Championship.
Notable Changes:
* Re-implementation of the container claw programming.
* An (untested) autonomous mode for hooking the container from the step.
* Commands for operating the container claw.
* Updated controls that allow the lift mechanism operator to slow the
descent of the lift clips.
* Updated PID algorithms for the new lift motors.
* DriveStraightToAutoZone now begins by lowering the lift mechanism to
lower the center of gravity for going over the scoring platform.
* A bug was fixed where the lift motors would jolt after the robot was
disabled, the motors rotated, and the robot re-enabled.

This might be the last commit to this repository (or at least for a very
long time), so goodbye to whoever out there might have been following
these commits and their messages. I wish the best for you and your
family. Good luck, and Godspeed with your journey through the rest of
your time on earth and thereafter.

Sincerely,
	Mackinnon Buck
	
I'll see you in the afterlife.